adapt comments fom review
