+++ assets/i18n/de.json
@@ -1658,7 +1658,7 @@
   "mayanumbers_single_numbers" : "Einzelne Zahlen",
   "mayanumbers_vigesimal" : "Zahl im Vigesimalsystem (Zahlenbasis 20)",
   "mayanumbers_mode_calendar" : "Kalender",
-  "mayanumbers_mode_calculate" : "Sonstig",
+  "mayanumbers_mode_calculate" : "Zahlen",
   "mexicanarmycipherwheel_title": "Mexican Army Chiffrierscheibe",
   "mexicanarmycipherwheel_description": "Verwendet von der Mexican Army im Krieg gegen die USA kurz vor dem 1. Weltkrieg",
   "mexicanarmycipherwheel_example": "TEST → 09580897",
