+++ assets/i18n/en.json
@@ -1658,7 +1658,7 @@
   "mayanumbers_single_numbers" : "Separate Numbers",
   "mayanumbers_vigesimal" : "Number in Vigesimal System (Numeral Base 20)",
   "mayanumbers_mode_calendar" : "Calendar",
-  "mayanumbers_mode_calculate" : "Other",
+  "mayanumbers_mode_calculate" : "Numbers",
   "mexicanarmycipherwheel_title": "Mexican Army Cipher Disk",
   "mexicanarmycipherwheel_description": "Used by the Mexican Army in a war with the USA before World War I.",
   "mexicanarmycipherwheel_example": "TEST → 09580897",
