+++ assets/i18n/fr.json
@@ -1518,7 +1518,7 @@
   "mayanumbers_single_numbers": "Nombres séparés",
   "mayanumbers_vigesimal": "Nombre dans le système Vigésimale (Base Numérique 20)",
   "mayanumbers_mode_calendar" : "Calendrier",
-  "mayanumbers_mode_calculate" : "Autre",
+  "mayanumbers_mode_calculate" : "Nombres",
   "mexicanarmycipherwheel_title": "Disque de cryptage de l'armée mexicaine",
   "mexicanarmycipherwheel_description": "Utilisé par l'armée mexicaine dans une guerre avec les États-Unis avant la Première Guerre mondiale.",
   "mexicanarmycipherwheel_example": "TEST → 09580897",
